@@ -1,6 +1,6 @@
 package ch.hevs.aislab.demo.database.entity;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.firebase.database.Exclude;
 
