@@ -1,7 +1,7 @@
 package ch.hevs.aislab.demo.ui.mgmt;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.Button;
