@@ -1,10 +1,10 @@
 package ch.hevs.aislab.demo.ui.transaction;
 
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProviders;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.view.GravityCompat;
+import androidx.annotation.NonNull;
+import androidx.core.view.GravityCompat;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
